Use hash tables to look up sessions and handles in the thread-safe session manager
(lwmsg: r42304)

+++ lwmsg/include/lwmsg/session.h
@@ -161,7 +161,7 @@ typedef enum LWMsgHandleType
 } LWMsgHandleType;
 
 #ifndef DOXYGEN
-typedef unsigned long LWMsgHandleID;
+typedef uint32_t LWMsgHandleID;
 
 typedef struct LWMsgSessionManager LWMsgSessionManager;
